# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: ytdl-service API
  version: 1.0.0
  description: |
    Downloads media with yt-dlp and serves it as MP4, MP3 or WAV files, live
    streams, or asynchronous tasks delivered by webhook.

    **Authentication.** Every `/v1` endpoint needs an API key, sent as
    `Authorization: Bearer ytdl_…` or `X-API-Key: ytdl_…`. Keys are issued by
    administrators and shown only once.

    **Roles.** Users can retrieve media, manage their own tasks, and read their
    own account and request history. Everything that changes data (users, API
    keys, catalog) and everything not owned by the caller needs the `admin`
    role.

    **Choosing the source.** Retrieval endpoints take `url` (a link) or `name`
    (a search query). When both are given, `url` wins.

    **Storage.** Retrieved files are stored and never downloaded twice. Files
    not requested for the configured idle period (default 7 days) are deleted
    and are downloaded again on the next request.

    **Idempotency.** `GET /v1/files` is idempotent. For `POST /v1/tasks`, an
    `Idempotency-Key` header makes a repeat return the same task. Without the
    header, an identical request returns the existing task while it is running
    or its result is still downloadable.

    **Rate limits** are per API key; exceeding them returns `429` with
    `Retry-After`. **Read responses** (lists, history, `/v1/me`) may be served
    from a short-lived cache, reported in `X-Cache`.

...
servers:
  - url: /
security:
  - bearerAuth: []
  - apiKeyHeader: []
...
paths:
...
  /openapi.yaml:
    get:
      tags: [Docs]
      operationId: openapiYAML
      summary: This specification (YAML)
      security: []
      responses:
        "200":
          description: The OpenAPI document.
          content:
            application/yaml:
              schema: { type: string }
        "304": { $ref: "#/components/responses/NotModified" }
  /openapi.json:
    get:
      tags: [Docs]
      operationId: openapiJSON
      summary: This specification (JSON)
      security: []
      responses:
        "200":
          description: The OpenAPI document.
          content:
            application/json:
              schema: { type: object }
        "304": { $ref: "#/components/responses/NotModified" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: ytdl_<43 base64url characters>
      description: "`Authorization: Bearer ytdl_…`"
    apiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
...
  headers:
...
    ETag:
      schema: { type: string }
...
  responses:
...
    NotModified:
      description: The resource matches `If-None-Match`; no body.
      headers:
        ETag: { $ref: "#/components/headers/ETag" }
...